blf変換/requirements.txt
python-can>=4.3
asammdf>=7.3
numpy>=1.24
cantools>=39.0
